Combined dish and recipe sections to one wrapper
